last update

+++ get next line/get_next_line_utils.c
@@ -6,15 +6,15 @@
 /*   By: otitebah <otitebah@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/18 10:56:44 by otitebah          #+#    #+#             */
-/*   Updated: 2022/11/25 22:10:17 by otitebah         ###   ########.fr       */
+/*   Updated: 2022/11/30 13:50:25 by otitebah         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "get_next_line.h"
 
-int	ft_strlen(char *str)
+size_t	ft_strlen(char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (!str)
